Added dtype=float32 parameter to get_fdata() because the default type of get_data() was float64, which caused tolerance issues

+++ clinica/iotools/converters/oasis_to_bids/oasis_to_bids.py
@@ -154,9 +154,9 @@ class OasisToBids(Converter):
             hdr["xyzt_units"] = 10
 
             img_with_good_orientation_nifti = nb.Nifti1Image(
-                np.round(img_with_wrong_orientation_analyze.get_fdata()).astype(
-                    np.int16
-                ),
+                np.round(
+                    img_with_wrong_orientation_analyze.get_fdata(dtype="float32")
+                ).astype(np.int16),
                 s_form,
                 header=hdr,
             )
